Removed standalone rotate_vector_collection.py module and migrated sole function into vector_utilities.py
